+++ src/models/Item.cairo
@@ -4,22 +4,22 @@ use starknet::ContractAddress;
 #[dojo::model]
 struct Item {
     #[key]
-    id: usize,
+    id: u32,
     name: felt252,
     // 1 - Melee Weapon, 2 - Ranged Weapon, 3 - Effect item, 4 - bag
     itemType: u8,
     // 0 - None, 1 - common, 2 - rare, 3 - legendary
     rarity: u8,
-    width: usize,
-    height: usize,
-    price: usize,
+    width: u32,
+    height: u32,
+    price: u32,
     // 0 - None, 1 - Damage, 2 - Cleanse Poison, 3 - Armor, 4 - Regen, 5 - Reflect, 6 - Poison, 7 - Empower, 8 - Vampirism, 9 - Expand pack
     effectType: u8,
     effectStacks: u32,
     // 0 - In armory, 1 - On Start, 2 - On Hit, 3 - On Cooldown, 4 - On Attack
     effectActivationType: u8,
     // Accuracy to trigger
-    chance: usize,
+    chance: u32,
     // item reuse time
     cooldown: u8,
     energyCost: u8,
@@ -31,5 +31,5 @@ struct Item {
 struct ItemsCounter {
     #[key]
     id: felt252,
-    count: usize,
+    count: u32,
 }
