+++ src/utils/random.cairo
@@ -13,7 +13,7 @@ fn pseudo_seed() -> (u128, u128, u128, u128) {
 }
 
 
-fn random(seed: u128, num: usize) -> usize {
+fn random(seed: u128, num: u32) -> u32 {
     let result = seed % num.into();
     result.try_into().unwrap()
 }
